Revert "Asymmetrical cookie signature - phase 1"
