+++ pan-domain-auth-core/src/main/scala/com/gu/pandomainauth/model/AuthenticatedUser.scala
