+++ pan-domain-auth-core/src/main/scala/com/gu/pandomainauth/model/User.scala
@@ -6,5 +6,4 @@ case class User(firstName: String, lastName: String, email: String, avatarUrl: O
   }
 
   def emailDomain = email.split("@").last
-  def username = email.split("@").head
 }
