+++ pan-domain-auth-core/src/main/scala/com/gu/pandomainauth/service/S3Bucket.scala
@@ -4,7 +4,6 @@ import java.util.Properties
 import com.amazonaws.ClientConfiguration
 import com.amazonaws.internal.StaticCredentialsProvider
 import com.amazonaws.regions.{Regions, Region}
-import com.gu.pandomainauth.PublicSettings
 
 import scala.collection.JavaConversions._
 import com.amazonaws.auth.AWSCredentials
@@ -19,7 +18,7 @@ class S3Bucket(credentials: Option[AWSCredentials] = None, regionOption: Option[
   val region = regionOption getOrElse(Region getRegion(Regions.EU_WEST_1))
   val s3Client = region.createClient(classOf[AmazonS3Client], credentialsProvider, awsClientConfiguration)
 
-  val bucketName = PublicSettings.bucketName
+  val bucketName = "pan-domain-auth-settings"
 
   def readDomainSettings(domain: String) = {
