+++ pan-domain-auth-example/conf/application.conf
@@ -62,6 +62,7 @@ host=${?HOST}
 aws {
   keyId = ${?AWS_ACCESS_KEY}
   secret = ${?AWS_SECRET}
+  bucket = ${?AWS_BUCKET}
 }
 
 include "local.conf"
