--- pan-domain-auth-verification/src/main/scala/com/gu/pandomainauth/model/AuthenticationStatus.scala
@@ -1,9 +1,0 @@
-package com.gu.pandomainauth.model
-
-sealed trait AuthenticationStatus
-case class Expired(authedUser: AuthenticatedUser) extends AuthenticationStatus
-case class GracePeriod(authedUser: AuthenticatedUser) extends AuthenticationStatus
-case class Authenticated(authedUser: AuthenticatedUser) extends AuthenticationStatus
-case class NotAuthorized(authedUser: AuthenticatedUser) extends AuthenticationStatus
-case class InvalidCookie(exception: Exception) extends AuthenticationStatus
-case object NotAuthenticated extends AuthenticationStatus
