--- pan-domain-auth-verification/src/main/scala/com/gu/pandomainauth/model/exceptions.scala
@@ -1,4 +1,0 @@
-package com.gu.pandomainauth.model
-
-class CookieSignatureInvalidException extends RuntimeException("cookie signature incorrect")
-class CookieParseException extends RuntimeException("cookie format incorrect")
